Refactored portsyncd structure to improve testability
